+++ lib/table_sync/event_actions/data_wrapper.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module TableSync::EventActions::DataWrapper
-  require_relative "data_wrapper/base"
-  require_relative "data_wrapper/destroy"
-  require_relative "data_wrapper/update"
-end