+++ lib/table_sync/publishing.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module TableSync
+  module Publishing
+    require_relative "publishing/base_publisher"
+    require_relative "publishing/publisher"
+    require_relative "publishing/batch_publisher"
+    require_relative "publishing/orm_adapter/active_record"
+    require_relative "publishing/orm_adapter/sequel"
+  end
+end