+++ lib/table_sync/utils.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module TableSync
+  module Utils
+    require_relative "utils/proc_array"
+    require_relative "utils/proc_keywords_resolver"
+    require_relative "utils/interface_checker"
+  end
+end