+++ spec/support/rails_settings.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class << Rails
-  def root
-    Pathname.new(__dir__).join("..", "..")
-  end
-end